Update tests, fix scoping issue and issue with expression-orientation
Because of the expression-oriented nature of Raven, every statement needs to return a value, but this value often comes from a block, like in the case of if statements. The block emits a value at the end, but still needs to close its scope and pop any local variables. The VM expects the stack to be empty at this point, but the return value of the block messes this up. A single-value reserve on the VM was added to remedy the problem.